LPS-60025 move method to the Base
